+++ src/generator/compute-model-params/compute-create-dto-params.ts
@@ -1,3 +1,5 @@
+import slash from 'slash';
+import path from 'node:path';
 import {
   DTO_CREATE_OPTIONAL,
   DTO_RELATION_CAN_CONNECT_ON_CREATE,
@@ -13,6 +15,7 @@ import {
   isReadOnly,
   isRelation,
   isRequiredWithDefaultValue,
+  isType,
   isUpdatedAt,
 } from '../field-classifiers';
 import {
@@ -20,6 +23,7 @@ import {
   concatUniqueIntoArray,
   generateRelationInput,
   getRelationScalars,
+  getRelativePath,
   makeImportsFromPrismaClient,
   mapDMMFToParsedField,
   zipImportStatementParams,
@@ -127,11 +131,51 @@ export const computeCreateDtoParams = ({
       overrides.isRequired = false;
     }
 
+    if (isType(field)) {
+      // don't try to import the class we're preparing params for
+      if (field.type !== model.name) {
+        const modelToImportFrom = allModels.find(
+          ({ name }) => name === field.type,
+        );
+
+        if (!modelToImportFrom)
+          throw new Error(
+            `related type '${field.type}' for '${model.name}.${field.name}' not found`,
+          );
+
+        const importName = templateHelpers.createDtoName(field.type);
+        const importFrom = slash(
+          `${getRelativePath(model.output.dto, modelToImportFrom.output.dto)}${
+            path.sep
+          }${templateHelpers.createDtoFilename(field.type)}`,
+        );
+
+        // don't double-import the same thing
+        // TODO should check for match on any import name ( - no matter where from)
+        if (
+          !imports.some(
+            (item) =>
+              Array.isArray(item.destruct) &&
+              item.destruct.includes(importName) &&
+              item.from === importFrom,
+          )
+        ) {
+          imports.push({
+            destruct: [importName],
+            from: importFrom,
+          });
+        }
+      }
+    }
+
     if (templateHelpers.config.classValidation) {
-      decorators.classValidators = parseClassValidators({
-        ...field,
-        ...overrides,
-      });
+      decorators.classValidators = parseClassValidators(
+        {
+          ...field,
+          ...overrides,
+        },
+        templateHelpers.createDtoName,
+      );
       concatUniqueIntoArray(
         decorators.classValidators,
         classValidators,
@@ -160,9 +204,18 @@ export const computeCreateDtoParams = ({
   }
 
   if (classValidators.length) {
+    if (classValidators.find((cv) => cv.name === 'Type')) {
+      imports.unshift({
+        from: 'class-transformer',
+        destruct: ['Type'],
+      });
+    }
     imports.unshift({
       from: 'class-validator',
-      destruct: classValidators.map((v) => v.name).sort(),
+      destruct: classValidators
+        .filter((cv) => cv.name !== 'Type')
+        .map((v) => v.name)
+        .sort(),
     });
   }